requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
jinja2==3.1.4
matplotlib==3.9.2
seaborn==0.13.2
pytest==8.3.3
httpx==0.27.2
xgboost==2.1.3

requests==2.32.3

# Optional (LLM target suggestion)
google-generativeai==0.8.3
